requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
python-multipart
python-multipart
edge-tts
requests

